feat(#539): moved core-lib
The content of the hanami_core-lib was moved into the main
binary, because for the upcoming task related hanges, the
cluster requires access to the database-tables and this was
the best solution.
