+++ Cargo.toml
@@ -1,7 +1,6 @@
 [workspace]
 members = [
     "src/hanami",
-    "src/libs/rust/hanami_core",
     "src/libs/rust/hanami_common",
     "src/libs/rust/hanami_cluster_parser",
     "src/libs/rust/hanami_dataset"
