+++ src/hanami/hanami_core_cpp
@@ -1,0 +1,1 @@
+../libs/cpp/hanami_core
