+++ src/hanami/src/api/http_endpoints/cluster/delete_cluster_v1_0.rs
@@ -20,9 +20,9 @@ use uuid::Uuid;
 use crate::api::errors::ErrorResponse;
 use crate::api::user_context::UserContext;
 use crate::database::cluster_table;
+use crate::core::cluster_handler;
 
 use hanami_common::enums;
-use hanami_core::cluster_handler;
 
 #[api_operation(
     tag = "cluster",
