--- src/libs/rust/hanami_core/.cargo/config.toml
@@ -1,3 +1,0 @@
-[build]
-rustflags = ["-C", "link-arg=-Wl,-rpath,$ORIGIN/../hanami_core_cpp"]
-
