--- src/libs/rust/hanami_core/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "hanami_core"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-hanami_common = { path = "../hanami_common" }
-hanami_dataset = { path = "../hanami_dataset" }
-hanami_cluster_parser = { path = "../hanami_cluster_parser" }
-
-# common stuff
-chrono = "0.4.40"
-uuid = { version = "1.16.0", features = ["v4", "serde"] }
-static_assertions = "1.1.0"
-lazy_static = "1.4"
-bytemuck = "1.22"
-
-# thread-binding
-core_affinity = "0.8"
-
-# json
-serde = { version = "1", features = ["derive", "serde_derive", "alloc"] }
-serde_with = "3.4"
-serde_json = "1"
-
-# logger
-log = "0.4"
-env_logger = "0.11"
-
-# c++
-cxx = "1.0"
-# Version 0.26 has much less warnings than, for example, 0.30, so this version is used for the moment
-autocxx = "0.26"
-
-[build-dependencies]
-autocxx-build = "0.26.0"
-cmake = "0.1"
