--- src/libs/rust/hanami_core/hanami_core_cpp
@@ -1,1 +1,0 @@
-../../cpp/hanami_core
