+++ src/libs/rust/hanami_dataset/Cargo.toml
@@ -10,7 +10,7 @@ serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 uuid = { version = "1.16.0", features = ["v4", "serde"] }
 # common stuff
-sanitize-filename = "0.5"
+sanitize-filename = "0.6"
 byteorder = "1.5"
 bincode = "1.3"
 bytemuck = "1.22"
